splits code into modules
